Fix weird auto scroll position when moving to the right/bottom of the sheet
In addition: add offset to the scroll_to_position function when iterating over search result items.

I forget why I clamped the scroll position in the first place. So this commit seems to "rollback" the any of the previous commit. I hope it won't causing regression. Well, may be it's a good idea to install the GitLens back?